+++ cmd/miau-desktop/frontend/src/lib/wailsjs/wailsjs/go/models.ts
@@ -483,6 +483,20 @@ export namespace desktop {
 	    }
 	}
 	
+	export class SyncResultDTO {
+	    newEmails: number;
+	    deletedEmails: number;
+	
+	    static createFrom(source: any = {}) {
+	        return new SyncResultDTO(source);
+	    }
+	
+	    constructor(source: any = {}) {
+	        if ('string' === typeof source) source = JSON.parse(source);
+	        this.newEmails = source["newEmails"];
+	        this.deletedEmails = source["deletedEmails"];
+	    }
+	}
 
 }